Some changes for Linux support

+++ src/Core/Application.cpp
@@ -24,7 +24,7 @@ namespace Core
 		return ApplicationCore::instance().LoadScene(p_sceneId);
 	}
 
-	__int64 Application::GetFrameCount()
+	long long Application::GetFrameCount()
 	{
 		return ApplicationCore::instance().GetFrameCount();
 	}
